181128|송태현|삭제 기능 수행시 product_id에 맞는 image 삭제 구현, product update 기능 구현

+++ src/main/java/shop/goodstudy/mall/image/model/Image.java
@@ -7,9 +7,9 @@ import lombok.Data;
 @Data
 @Alias("image")
 public class Image {
-	private int image_id;
+	private long image_id;
 	private String physical_name;
 	private String image_name;
 	private byte[] imagefile;
-	private int product_id;
+	private long product_id;
 }
